+++ src/components/hooks/useDevices.ts
@@ -1,11 +1,16 @@
-import { DependencyList, useContext } from "react";
-import { DeviceFilter } from "../../../jacdac-ts/src/jdom/bus";
-import JacdacContext, { JacdacContextProps } from "../../jacdac/Context";
-import useChange from '../../jacdac/useChange';
+import { DependencyList, useContext } from "react"
+import { DeviceFilter } from "../../../jacdac-ts/src/jdom/bus"
+import JacdacContext, { JacdacContextProps } from "../../jacdac/Context"
+import useChange from "../../jacdac/useChange"
 
-export default function useDevices(options?: DeviceFilter, deps: DependencyList = []) {
+export default function useDevices(
+    options?: DeviceFilter,
+    deps: DependencyList = []
+) {
     const { bus } = useContext<JacdacContextProps>(JacdacContext)
-    const devices = useChange(bus, b => b.devices(options)
-        , [JSON.stringify(options), ...deps])
-    return devices;
+    const devices = useChange(bus, _ => _?.devices(options) || [], [
+        JSON.stringify(options),
+        ...deps,
+    ])
+    return devices
 }
