+++ src/pages/tools/vm-editor.tsx
@@ -1,264 +1,17 @@
-import { createStyles, makeStyles } from "@material-ui/core"
+import { Grid } from "@material-ui/core"
 import React from "react"
-import ReactBlockly from "react-blockly"
-import Blockly from "blockly"
-import Theme from "@blockly/theme-modern"
+import VmEditor from "../../components/blockly/VmEditor"
+import Dashboard from "../../components/dashboard/Dashboard"
 
-const useStyles = makeStyles(() =>
-    createStyles({
-        editor: {
-            height: "calc(100vh - 10rem)",
-        },
-    })
-)
-
-const initialXml =
-    '<xml xmlns="http://www.w3.org/1999/xhtml"><block type="jacdac_configuration"></block></xml>'
-const blocks = [
-    {
-        type: "jacdac_configuration",
-        message0: "configuration",
-        inputsInline: true,
-        nextStatement: "Role",
-        style: "variable_blocks",
-    },
-    {
-        type: "jacdac_declare_role",
-        message0: "define role %1 as %2",
-        args0: [
-            {
-                type: "field_variable",
-                name: "NAME",
-                variable: "button 1",
-                defaultType: "Service",
-            },
-            {
-                type: "field_dropdown",
-                name: "SERVICE",
-                options: [
-                    ["button", "button"],
-                    ["servo", "servo"],
-                    ["settings", "settings"],
-                ],
-            },
-        ],
-        style: "variable_blocks",
-        previousStatement: "Role",
-        nextStatement: "Role",
-    },
-    {
-        type: "jacdac_button_event",
-        message0: "when %1 %2 %3",
-        args0: [
-            {
-                type: "field_variable",
-                name: "ROLE",
-                variable: "button",
-            },
-            {
-                type: "input_dummy",
-            },
-            {
-                type: "field_dropdown",
-                name: "EVENT",
-                options: [
-                    ["down", "down"],
-                    ["up", "up"],
-                    ["press", "press"],
-                ],
-            },
-        ],
-        style: "logic_blocks",
-        inputsInline: true,
-        nextStatement: null,
-        tooltip: "",
-        helpUrl: "",
-    },
-    {
-        type: "jacdac_reading_change",
-        message0: "when %1 change",
-        args0: [
-            {
-                type: "field_variable",
-                name: "ROLE",
-                variable: "humidity",
-            },
-        ],
-        inputsInline: true,
-        nextStatement: "Statement",
-        style: "logic_blocks",
-        tooltip: "",
-        helpUrl: "",
-    },
-    {
-        type: "jacdac_await_condition",
-        message0: "while %1",
-        args0: [
-            {
-                type: "input_value",
-                name: "CONDITION",
-                check: "Boolean",
-            },
-        ],
-        style: "logic_blocks",
-        inputsInline: true,
-        nextStatement: "Statement",
-        tooltip: "",
-        helpUrl: "",
-    },
-    {
-        type: "jacdac_wait_ms",
-        message0: "wait %1 ms",
-        args0: [
-            {
-                type: "field_number",
-                name: "NAME",
-                value: 0,
-                min: 0,
-                max: 5000,
-            },
-        ],
-        inputsInline: true,
-        previousStatement: "Statement",
-        nextStatement: "Statement",
-        colour: 230,
-        tooltip: "",
-        helpUrl: "",
-    },
-    {
-        type: "jacdac_button_reading",
-        message0: "%1 pressure",
-        args0: [
-            {
-                type: "field_variable",
-                name: "ROLE",
-                variable: "button",
-            },
-        ],
-        inputsInline: true,
-        output: "Number",
-        colour: 230,
-        tooltip: "",
-        helpUrl: "",
-    },
-    {
-        type: "jacdac_humidity_reading",
-        message0: "%1 humidity",
-        args0: [
-            {
-                type: "field_variable",
-                name: "ROLE",
-                variable: "humidity",
-            },
-        ],
-        inputsInline: true,
-        output: "Number",
-        colour: 230,
-        tooltip: "",
-        helpUrl: "",
-    },
-    {
-        type: "jacdac_servo_write_angle",
-        message0: "set %1 angle to %2 %3",
-        args0: [
-            {
-                type: "field_variable",
-                name: "ROLE",
-                variable: "servo",
-            },
-            {
-                type: "input_dummy",
-            },
-            {
-                type: "input_value",
-                name: "VALUE",
-                check: "Number",
-            },
-        ],
-        inputsInline: true,
-        previousStatement: "Statement",
-        nextStatement: "Statement",
-        colour: 230,
-        tooltip: "",
-        helpUrl: "",
-    },
-    {
-        type: "jacdac_servo_enable",
-        message0: "set %1 %2",
-        args0: [
-            {
-                type: "field_variable",
-                name: "ROLE",
-                variable: "servo",
-            },
-            {
-                type: "field_dropdown",
-                name: "ENABLED",
-                options: [
-                    ["on", "ON"],
-                    ["off", "OFF"],
-                ],
-            },
-        ],
-        colour: 230,
-        tooltip: "",
-        helpUrl: "",
-        previousStatement: "Statement",
-        nextStatement: "Statement",
-    },
-]
-blocks.map(
-    block =>
-        (Blockly.Blocks[block.type] = {
-            init: function () {
-                this.jsonInit(block)
-            },
-        })
-)
-const toolboxBlocks = [
-    ...blocks.map(block => ({ type: block.type })),
-    { type: "math_number" },
-    { type: "math_arithmetic" },
-    { type: "logic_boolean" },
-    { type: "logic_compare" },
-    { type: "logic_operation" },
-    { type: "logic_negate" },
-]
-
-export default function VMEditor() {
-    const classes = useStyles()
-    const handleChange = (workspace: Blockly.WorkspaceSvg) => {
-        const newXml = Blockly.Xml.domToText(
-            Blockly.Xml.workspaceToDom(workspace)
-        )
-        document.getElementById("generated-xml").innerText = newXml
-
-        // const code = Blockly.JavaScript.workspaceToCode(workspace)
-        //  document.getElementById("code").value = code
-    }
-
+export default function Page() {
     return (
-        <>
-            <ReactBlockly
-                toolboxBlocks={toolboxBlocks}
-                workspaceConfiguration={{
-                    comments: false,
-                    css: true,
-                    grid: {
-                        spacing: 25,
-                        length: 1,
-                        colour: "#ccc",
-                        snap: true,
-                    },
-                    renderer: "zelos",
-                    theme: Theme,
-                    oneBasedIndex: false,
-                }}
-                initialXml={initialXml}
-                wrapperDivClassName={classes.editor}
-                workspaceDidChange={handleChange}
-            />
-            <pre id="generated-xml"></pre>
-        </>
+        <Grid container spacing={1}>
+            <Grid item xs={12} sm={5} lg={4} xl={3}>
+                <Dashboard />
+            </Grid>
+            <Grid item xs={12} sm={7} lg={8} xl={9}>
+                <VmEditor />
+            </Grid>
+        </Grid>
     )
 }
